requirements.txt
absl-py==2.1.0
accelerate==0.29.3
antlr4-python3-runtime==4.9.3
attrs==23.2.0
bitsandbytes==0.43.1
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==4.4.2
faster-fifo==1.4.5
filelock==3.13.4
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
gym==0.23.0
gym-notices==0.0.8
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==3.6
ifcfg==0.24
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.59.1
numpy==1.23.5
numpy-quaternion==2023.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objectio==0.2.29
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
proglog==0.1.10
protobuf==3.20.1
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.0.4
pygame==2.0.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scipy==1.13.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
webdataset==0.1.40
Werkzeug==3.0.2
zipp==3.18.1
